Modified to start with Apache, Added Validator and Pagination

+++ src/Application/DeleteUser/DeleteUserService.php
@@ -7,6 +7,8 @@ use App\Domain\User\UserId;
 
 class DeleteUserService
 {
+    private IUserRepository $userRepository;
+
     public function __construct(IUserRepository $userRepository)
     {
         $this->userRepository = $userRepository;
@@ -15,6 +17,12 @@ class DeleteUserService
     public function execute(DeleteUserRequest $request): void
     {
         $id = new UserId($request->id());
+
+        $user = $this->userRepository->findById($id);
+
+        if(!$user){
+            throw new UserDoesNotExistException();
+        }
 
         $this->userRepository->deleteById($id);
     }
